egoscale_research/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "egoscale-research"
version = "0.1.0"
description = "Spec-first EgoScale research scaffold."
requires-python = ">=3.9"
dependencies = [
  "torch>=2.2",
  "transformers>=4.57.0",
  "PyYAML>=6.0",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
